refactor nn.Linear
